v2

+++ src/views/auth/RegisterView.vue
@@ -2,31 +2,57 @@
 import { ref } from "vue";
 import { useAuthStore } from "@/stores/auth";
 import { useRouter } from "vue-router";
+import BaseAlertError from "@/components/Alert/BaseAlertError.vue";
+import { watch } from "vue";
 
 const authStore = useAuthStore();
-const { userLogin } = authStore;
+const { registerUser } = authStore;
 
+const error = ref(false)
 const router = useRouter();
+const disabled = ref(true)
 
-const email = ref(null);
-const password = ref(null);
+const passwordConfirm = ref('')
+const payload = ref({
+  name: '',
+  email: '',
+  password: '',
+});
 
 const handleSubmit = async () => {
-  const payload = {
-    email: email.value,
-    password: password.value,
-  };
+  if(payload.value.password !== passwordConfirm.value){
+    error.value = true 
+    setTimeout(() => {
+      error.value = false
+    }, 3000)
+    return 
+  }
 
-  const login = await userLogin(payload);
-  console.log("a", login);
+  const login = await registerUser(payload.value);
 
   if (login) {
     localStorage.setItem("token-auth", login.token);
-    localStorage.setItem("user_id", login.user_id);
+    localStorage.setItem("user_id", login.user.id);
 
     router.push({ path: "/home" });
+  } else {
+    error.value = true
+
+    setTimeout(() => {
+      error.value = false
+    }, 3000)
   }
 };
+
+watch(payload.value, () => {
+  if(payload.value.email.length > 1 && payload.value.password.length > 1  && payload.value.name.length > 1){
+    disabled.value = false
+  } else {
+    disabled.value = true
+  }
+
+ 
+})
 </script>
 
 <template>
@@ -34,35 +60,63 @@ const handleSubmit = async () => {
     <div class="flex column form-container">
       <h1>Cadastre-se</h1>
       <div class="flex column input-container">
-        <label>Email</label>
+        <label>Nome de Usuário</label>
         <input
           type="text"
+          placeholder="Insira seu nome de usuário"
+          :value="payload.name"
+          @input="(ev) => (payload.name = ev.target.value)"
+        />
+      </div>
+      <div class="flex column input-container">
+        <label>Email</label>
+        <input
+          type="email"
           placeholder="Insira seu email"
-          :value="email"
-          @input="(ev) => (email = ev.target.value)"
+          :value="payload.email"
+          @input="(ev) => (payload.email = ev.target.value)"
         />
       </div>
       <div class="flex column input-container">
         <label>Senha</label>
         <input
-          type="text"
-          placeholder="Insira seu email"
-          :value="password"
-          @input="(ev) => (password = ev.target.value)"
+          type="password"
+          placeholder="Crie sua senha"
+          :value="payload.password"
+          @input="(ev) => (payload.password = ev.target.value)"
         />
       </div>
-      <button @click="handleSubmit">Criar conta</button>
+      <div class="flex column input-container">
+        <label>Confirme sua senha</label>
+        <input
+          type="password"
+          placeholder="Crie sua senha"
+          :value="passwordConfirm"
+          @input="(ev) => (passwordConfirm = ev.target.value)"
+        />
+      </div>
+
+    
+
+      <button @click="handleSubmit" :class="disabled === true ? 'disabled' : ''" :disabled="disabled">Entrar</button>
       <span>
-        Já possui uma conta?
+        Tem uma conta?
         <RouterLink class="login-span-text" to="/login"
           >Faça seu login</RouterLink
         >
       </span>
+      <BaseAlertError v-if="error" type="error" text="As senhas não coincidem"/>
+
     </div>
   </div>
 </template>
 
 <style scoped lang="scss">
+.disabled{
+  opacity: 0.6;
+ 
+}
+
 .password-span-text {
   color: blue;
   text-align: end;
@@ -102,6 +156,7 @@ const handleSubmit = async () => {
     border: 1px solid rgba(83, 83, 83, 0.219);
     border-radius: 8px;
     font-size: 1rem;
+    cursor: pointer;
   }
 }
 @media (min-width: 0px) and (max-width: 768px) {
